+++ src/pxfquery/engines/__init__.py
@@ -1,6 +1,0 @@
-from pxfquery.engines.forward_drug import run as run_forward_drug
-from pxfquery.engines.forward_genetic import run as run_forward_genetic
-from pxfquery.engines.reverse_drug import run as run_reverse_drug
-from pxfquery.engines.reverse_genetic import run as run_reverse_genetic
-
-__all__ = ["run_forward_drug", "run_forward_genetic", "run_reverse_drug", "run_reverse_genetic"]